refactor: rename installing package to install
