+++ internal/install/uninstall.go
@@ -1,4 +1,4 @@
-package installing
+package install
 
 import (
 	"fmt"
